Converted Phactory to use PHP namespaces and instance objects
The structure of Phactory was causing problems with autoloaders, since
there were Phactory.php and a PhactoryMongo.php files, but each file
contained a class with the same name. In addition, the classes for
MongoDB were encapsulated in a PhactoryMongo subdirectory, but the
classes for testing SQL databases were not.

Furthermore, the use of a fully static class caused fatal errors when
attempting to run the unit tests with PHPUnit. This is because of the
static property for $_pdo. PHPUnit was attempting to serialize this
object and failing to do so. Unfortunately, efforts to set
backupStaticAttributes to false in phpunit.xml or disable it at the
class/method level failed, and I found that this could also be related
to the following bug:

https://github.com/sebastianbergmann/phpunit/issues/254

After setting processIsolation="false", stopOnFailure="false", and
stopOnError="false", I was able to debug indiviual problems one-by-one
until the tests all passed, then I could turn processIsolation back on,
and the tests continued to pass.

As a result of my changes, this version of Phactory BREAKS BACKWARDS
COMPATIBILITY with all previous versions, but I think it leads to a
cleaner, easier-to-use, and easier-to-test interface.